Minor edits
Just a quick look at the abstract, will continue later.

I have a concern about the abstract in that it doesn't discuss the responses of seagrass or macro-algae, which didn't seem to be abnormal (seagrass) or too abnormal (macro). As written the abstract states that everything had abnormal results.

I'll try to make the change wherever I see it, but 'spp.' (as in *Dapis* spp.) should not be italicized.